+++ src/basemode/basemode.cpp
@@ -34,32 +34,32 @@ This file is part of Liberal Crime Squad.                                       
 
 
 void mode_base(void) {
-    short advanced = 0;
-    short directing = 0;
-    short in_newspaper = 0;
-    short in_halloween = 0;
-    short in_halloween2 = 0;
-    short in_training = 0;
-    short in_compound = 0;
+    int16 advanced = 0;
+    int16 directing = 0;
+    int16 in_newspaper = 0;
+    int16 in_halloween = 0;
+    int16 in_halloween2 = 0;
+    int16 in_training = 0;
+    int16 in_compound = 0;
 
-    short investing = 0;
-    short investing_newspaper = 0;
-    short investing_halloween = 0;
-    short investing_training = 0;
-    short investing_compound = 0;
-    short investing_stores = 0;
+    int16 investing = 0;
+    int16 investing_newspaper = 0;
+    int16 investing_halloween = 0;
+    int16 investing_training = 0;
+    int16 investing_compound = 0;
+    int16 investing_stores = 0;
 
-    short buyer = 0;
+    int16 buyer = 0;
 
     char forcewait, canseethings;
-    long nonsighttime = 0;
+    int32 nonsighttime = 0;
     #ifdef SHOWWAIT
-    int oldforcemonth = month;
+    int32 oldforcemonth = month;
     #endif
 
-    int length = 0;
+    int32 length = 0;
 
-    long l = 0;
+    int32 l = 0;
 
     do {
         forcewait = 1;
@@ -67,7 +67,7 @@ void mode_base(void) {
         cantseereason = 3;
 
         if(!disbanding) {
-            for(int p = 0; p < pool.size(); p++) {
+            for(int32 p = 0; p < pool.size(); p++) {
                 if(pool[p]->alive &&
                         pool[p]->align == 1 &&
                         pool[p]->dating == 0 &&
@@ -105,9 +105,9 @@ void mode_base(void) {
             addstr(num);
 
 
-            int y = 2;
+            int32 y = 2;
 
-            for(int v = 0; v < VIEWNUM; v++) {
+            for(int32 v = 0; v < VIEWNUM; v++) {
                 if(attitude[VIEW_LIBERALCRIMESQUAD] == 0 &&
                         v == VIEW_LIBERALCRIMESQUADPOS)
                     continue;
@@ -250,11 +250,11 @@ void mode_base(void) {
             nonsighttime = 0;
         }
 
-        int partysize = 0;
-        int partydead = 0;
+        int32 partysize = 0;
+        int32 partydead = 0;
 
         if(activesquad != NULL) {
-            for(int p = 0; p < 6; p++) {
+            for(int32 p = 0; p < 6; p++) {
                 if(activesquad->squad[p] != NULL)
                     partysize++;
                 else {
@@ -275,7 +275,7 @@ void mode_base(void) {
             }
         }
 
-        long safenumber = 0;
+        int32 safenumber = 0;
 
         for(l = 0; l < location.size(); l++) {
             if(location[l]->renting >= 0)
@@ -364,7 +364,7 @@ void mode_base(void) {
             }
 
             if(haveflag) {
-                for(int p = 0; p < 7; p++) {
+                for(int32 p = 0; p < 7; p++) {
                     move(p + 17, 32);
 
                     if(p < 3) {
@@ -374,7 +374,7 @@ void mode_base(void) {
                         set_color(COLOR_WHITE, COLOR_RED, 1);
                         move(p + 17, 38);
 
-                        for(int i = 0; i < 10; i++)
+                        for(int32 i = 0; i < 10; i++)
                             addch(CH_LOWER_HALF_BLOCK);
                     } else {
                         if(p < 6)
@@ -382,7 +382,7 @@ void mode_base(void) {
                         else
                             set_color(COLOR_RED, COLOR_BLACK, 0);
 
-                        for(int i = 0; i < 16; i++) {
+                        for(int32 i = 0; i < 16; i++) {
                             if(p == 6)
                                 addch(CH_UPPER_HALF_BLOCK);
                             else
@@ -544,7 +544,7 @@ void mode_base(void) {
             refresh();
         }
 
-        int c = 'w';
+        int32 c = 'w';
 
         if(!forcewait) {
             c = getch();
@@ -592,7 +592,7 @@ void mode_base(void) {
             if(activesquad == NULL)
                 activesquad = squad[0];
             else {
-                for(int sq = 0; sq < squad.size(); sq++) {
+                for(int32 sq = 0; sq < squad.size(); sq++) {
                     if(squad[sq] == activesquad) {
                         if(sq == squad.size() - 1)
                             activesquad = squad[0];
@@ -607,14 +607,14 @@ void mode_base(void) {
 
         if(c == 'z' && safenumber > 0) {
             activesquad = NULL;
-            long sl;
+            int32 sl;
 
             if(selectedsiege == -1)
                 sl = 0;
             else
                 sl = selectedsiege + 1;
 
-            for(long l = sl; l < location.size(); l++) {
+            for(int32 l = sl; l < location.size(); l++) {
                 if(location[l]->renting >= 0) {
                     selectedsiege = l;
                     break;
