+++ src/common/equipment.cpp
@@ -31,7 +31,7 @@ This file is part of Liberal Crime Squad.                                       
 
 
 /* review squad equipment */
-void equip(vector<itemst *> &loot, int loc) {
+void equip(vector<itemst *> &loot, int32 loc) {
     if(activesquad == NULL)
         return;
 
@@ -40,7 +40,7 @@ void equip(vector<itemst *> &loot, int loc) {
     if(loc != -1)
         consolidateloot(location[loc]->loot);
 
-    int page = 0;
+    int32 page = 0;
 
     do {
         erase();
@@ -51,10 +51,10 @@ void equip(vector<itemst *> &loot, int loc) {
 
         printparty();
 
-        int x = 1, y = 10;
+        int32 x = 1, y = 10;
         char str[200], str2[200];
 
-        for(int l = page * 18; l < loot.size() && l < page * 18 + 18; l++) {
+        for(int32 l = page * 18; l < loot.size() && l < page * 18 + 18; l++) {
             if(loot[l]->type == ITEM_WEAPON) {
                 getweaponfull(str2, loot[l]->weapon.type);
 
@@ -165,11 +165,11 @@ void equip(vector<itemst *> &loot, int loc) {
 
         refresh();
 
-        int c = getch();
+        int32 c = getch();
         translategetch(c);
 
         if(c >= 'a' && c <= 'r') {
-            int slot = c - 'a' + page * 18;
+            int32 slot = c - 'a' + page * 18;
 
             if(slot >= 0 && slot < loot.size()) {
                 move(8, 20);
@@ -178,12 +178,12 @@ void equip(vector<itemst *> &loot, int loc) {
 
                 refresh();
 
-                int c = getch();
+                int32 c = getch();
                 translategetch(c);
 
                 if(c >= '1' && c <= '6') {
                     if(activesquad->squad[c - '1'] != NULL) {
-                        int armok = 2;
+                        int32 armok = 2;
 
                         if((activesquad->squad[c - '1']->wound[BODYPART_ARM_RIGHT] & WOUND_NASTYOFF) ||
                                 (activesquad->squad[c - '1']->wound[BODYPART_ARM_RIGHT] & WOUND_CLEANOFF))
@@ -217,11 +217,11 @@ void equip(vector<itemst *> &loot, int loc) {
                             }
 
                             //DROP ALL CLIPS THAT DON'T WORK
-                            for(int cl = 0; cl < CLIPNUM; cl++) {
+                            for(int32 cl = 0; cl < CLIPNUM; cl++) {
                                 if(cl == ammotype(activesquad->squad[c - '1']->weapon.type))
                                     continue;
 
-                                for(int p2 = 0; p2 < activesquad->squad[c - '1']->clip[cl]; p2++) {
+                                for(int32 p2 = 0; p2 < activesquad->squad[c - '1']->clip[cl]; p2++) {
                                     itemst *newi = new itemst;
                                     newi->type = ITEM_CLIP;
                                     newi->cliptype = cl;
@@ -282,7 +282,7 @@ void equip(vector<itemst *> &loot, int loc) {
 
             refresh();
 
-            int c = getch();
+            int32 c = getch();
             translategetch(c);
 
             if(c >= '1' && c <= '6') {
@@ -315,7 +315,7 @@ void equip(vector<itemst *> &loot, int loc) {
         }
 
         if(c >= '1' && c <= '6') {
-            int p = c - '1';
+            int32 p = c - '1';
 
             if(activesquad->squad[p] != NULL) {
                 if(activesquad->squad[p]->weapon.type != WEAPON_NONE) {
@@ -328,8 +328,8 @@ void equip(vector<itemst *> &loot, int loc) {
                     activesquad->squad[p]->weapon.ammo = 0;
 
                     //DROP ALL CLIPS
-                    for(int c = 0; c < CLIPNUM; c++) {
-                        for(int p2 = 0; p2 < activesquad->squad[p]->clip[c]; p2++) {
+                    for(int32 c = 0; c < CLIPNUM; c++) {
+                        for(int32 p2 = 0; p2 < activesquad->squad[p]->clip[c]; p2++) {
                             itemst *newi = new itemst;
                             newi->type = ITEM_CLIP;
                             newi->cliptype = c;
@@ -359,14 +359,14 @@ void equip(vector<itemst *> &loot, int loc) {
 
 /* lets you pick stuff to stash/retrieve from one location to another */
 void moveloot(vector<itemst *> &dest, vector<itemst *> &source) {
-    unsigned long ret = 0;
+    uint32 ret = 0;
 
-    int page = 0;
+    int32 page = 0;
 
     vector<char> selected;
     selected.resize(source.size());
 
-    for(int s = 0; s < selected.size(); s++)
+    for(int32 s = 0; s < selected.size(); s++)
         selected[s] = 0;
 
     do {
@@ -378,10 +378,10 @@ void moveloot(vector<itemst *> &dest, vector<itemst *> &source) {
 
         printparty();
 
-        int x = 1, y = 10;
+        int32 x = 1, y = 10;
         char str[200], str2[200];
 
-        for(int l = page * 18; l < source.size() && l < page * 18 + 18; l++) {
+        for(int32 l = page * 18; l < source.size() && l < page * 18 + 18; l++) {
             if(selected[l])
                 set_color(COLOR_GREEN, COLOR_BLACK, 1);
             else
@@ -473,11 +473,11 @@ void moveloot(vector<itemst *> &dest, vector<itemst *> &source) {
 
         refresh();
 
-        int c = getch();
+        int32 c = getch();
         translategetch(c);
 
         if(c >= 'a' && c <= 'r') {
-            int slot = c - 'a' + page * 18;
+            int32 slot = c - 'a' + page * 18;
 
             if(slot >= 0 && slot < source.size()) {
                 if(selected[slot])
@@ -531,7 +531,7 @@ void moveloot(vector<itemst *> &dest, vector<itemst *> &source) {
 
     } while(1);
 
-    for(int l = source.size() - 1; l >= 0; l--) {
+    for(int32 l = source.size() - 1; l >= 0; l--) {
         if(selected[l] > 0) {
             itemst *newit = new itemst;
             *newit = *source[l];
@@ -552,7 +552,7 @@ void moveloot(vector<itemst *> &dest, vector<itemst *> &source) {
 
 /* combines multiple items of the same type into stacks */
 void consolidateloot(vector<itemst *> &loot) {
-    int l, l2;
+    int32 l, l2;
 
     //PUT THINGS TOGETHER
     for(l = loot.size() - 1; l >= 1; l--) {
@@ -705,7 +705,7 @@ char itemcompare(itemst *a, itemst *b) {
 
 
 /* returns the type of ammo used by the given weapon, if any */
-short ammotype(int type) {
+int16 ammotype(int32 type) {
     switch(type) {
     case WEAPON_REVOLVER_22:
         return CLIP_22;
@@ -759,8 +759,8 @@ char rangedweapon(weaponst &w) {
 
 
 /* check if the squad has a certain weapon */
-char squadhasitem(squadst &sq, int type, int subtype) {
-    for(int p = 0; p < 6; p++) {
+char squadhasitem(squadst &sq, int32 type, int32 subtype) {
+    for(int32 p = 0; p < 6; p++) {
         if(sq.squad[p] != NULL) {
             switch(type) {
             case ITEM_WEAPON:
@@ -772,7 +772,7 @@ char squadhasitem(squadst &sq, int type, int subtype) {
         }
     }
 
-    for(int l = 0; l < sq.loot.size(); l++) {
+    for(int32 l = 0; l < sq.loot.size(); l++) {
         if(sq.loot[l]->type != type)
             continue;
